#14 restructure files
