+++ src/pages/Tasks/components/Edit/TaskCompleteToggle.tsx
@@ -2,7 +2,7 @@ import classNames from 'classnames';
 import { useTranslation } from 'react-i18next';
 
 import { BaseComponentProps } from 'common/utils/types';
-import { Task } from 'pages/Users/api/useGetUserTasks';
+import { Task } from 'pages/Tasks/api/useGetUserTasks';
 import { useUpdateTask } from 'pages/Tasks/api/useUpdateTask';
 import { useToasts } from 'common/hooks/useToasts';
 import FAIcon from 'common/components/Icon/FAIcon';
