+++ src/pages/Users/components/UserTaskListItem.tsx
@@ -1,5 +1,5 @@
 import { BaseComponentProps } from 'common/utils/types';
-import { Task } from '../api/useGetUserTasks';
+import { Task } from '../../Tasks/api/useGetUserTasks';
 import TaskCompleteToggle from '../../Tasks/components/Edit/TaskCompleteToggle';
 import Link from 'common/components/Link/Link';
