+++ src/pages/Users/components/__tests__/UserList.test.tsx
@@ -2,7 +2,7 @@ import { describe, expect, it, vi } from 'vitest';
 import { render, screen } from 'test/test-utils';
 
 import * as UseGetUser from 'common/api/useGetUser';
-import * as UseGetUsers from '../../api/useGetUsers';
+import * as UseGetUsers from '../../../Tasks/api/useGetUsers';
 import UserList from '../UserList';
 import { UseQueryResult } from '@tanstack/react-query';
 import { AxiosError } from 'axios';
