+++ src/pages/Users/components/__tests__/UserTaskList.test.tsx
@@ -1,7 +1,7 @@
 import { describe, expect, it, vi } from 'vitest';
 import { render, screen } from 'test/test-utils';
 
-import * as UseGetUserTasks from '../../api/useGetUserTasks';
+import * as UseGetUserTasks from '../../../Tasks/api/useGetUserTasks';
 import UserTaskList from '../UserTaskList';
 import { UseQueryResult } from '@tanstack/react-query';
 import { AxiosError } from 'axios';
